+++ genome_comparator/__main__.py
@@ -1,0 +1,6 @@
+"""Allow running the main command with "python -m genome_comparator"."""
+
+from .cli import main
+
+if __name__ == '__main__':
+    main()